requirements-doc.txt
sphinx>=3.4.3,<4
sphinx-autodoc-typehints>=1.11.1
sphinx-rtd-theme>=0.5.1,<1
autodocsumm>=0.2.2,<1
sphinx-icontract>=2.0.2
